Comments added & Refactorization
While beginning to add scoring functionality to this model, discovered that there were overlaps in some areas of the DNA geometry. This was due to the fact that the cylindrical fibre volume used by the nuclear fractal arrangement was different (smaller in radius & length) than the actual fibre model (from DnaFabric). To fix this:
- Fractal cylinder radius increased 12 —> 17 nm
- DnaFabric cylinder length decreased 160 —> 136 nm (= 170 - 17(2))
- Fibre pitch decreased  10 —> 8.5 nm (= 136/160 x 10)
These changes were verified in terms of clearance of overlaps & fit within the nuclear ellipse volume.

Additionally, the geometry code was extensively commented to improve readability & to better understand how to access the various logical volumes of the DNA elements in order to perform scoring within them.