Modified parameter file and date files.
Objective of this commit was to allow user to vary the total dose threshold and still use the relative dose contributions of secondary particle species. When reading dose values from data files, previously the dose was normalized to 1 Gy. Now the data files contained an unsigned double corresponding to their relative dose contribution. In the main parameter file, this value is used to scale a total target dose value (over all secondary species) to the appropriate dose threshold for this species.